[Chore] MVVM 아키텍처 적용
