+++ MalatangChallenge/Model/Malatang.mlmodel
