+++ MalatangChallenge/View/LoadingView.swift
