+++ MalatangChallenge/View/ReadyView.swift
