+++ MalatangChallenge/View/ResultView.swift
